attempting to assign the class of the QScrollArea to the result of a function which should eventually explicitly calculate and set the height. This appears to be required to make CSS happy.
The general structure is QScrollArea->Inner Div->Content Container DIV->Content Div's ... It's the Inner Div whose height is 0. The height of the Content Container DIV is correct and seems to be calculated based on the total height of the Content Div's. When the height of the QScrollArea is explicitly set, the Inner Div gets the same height at the QScrollArea. When the height of the QScrollArea is calculated automatically, the Inner Div does not receive the height of the QScrollArea. I am guessing that in the creation of the Inner Div by the QScrollArea, the height of the inner div is not updated after the height of the QScrollArea is calculated.